refactor: add test in-memory database for remeaning targets

+++ composeApp/src/desktopMain/kotlin/br/com/noartcode/theprice/domain/usecases/ICurrencyFormatter.desktop.kt
@@ -2,11 +2,11 @@ package br.com.noartcode.theprice.domain.usecases
 
 actual class CurrencyFormatter : ICurrencyFormatter {
 
-    override fun format(value: Int): String {
+    override fun format(value: Long): String {
         TODO("Not yet implemented")
     }
 
-    override fun clenup(value: String): Int {
+    override fun clenup(value: String): Long {
         TODO("Not yet implemented")
     }
 }
